src/requirements.txt
fastapi
uvicorn[standard]
yt-dlp[curl-cffi]
yt-dlp-ejs
requests
bilibili-api-python
aiohttp
python-dotenv
python-multipart